+++ src/console/controllers/MigrateController.php
@@ -230,8 +230,9 @@ class MigrateController extends BaseMigrateController
 
         // TODO remove after next breakpoint
         // Make sure that the new project config structure is there before any migrations cause Project Config to look there.
-        if (!file_exists(Craft::$app->getPath()->getProjectConfigFilePath())) {
-            Craft::$app->getProjectConfig()->regenerateYamlFromConfig();
+        $projectConfig = Craft::$app->getProjectConfig();
+        if ($projectConfig->writeYamlAutomatically && !$projectConfig->getDoesYamlExist()) {
+            $projectConfig->regenerateYamlFromConfig();
         }
 
         if (!parent::beforeAction($action)) {