+++ src/helpers/Console.php
@@ -125,14 +125,15 @@ class Console extends \yii\helpers\Console
     }
 
     /**
-     * Ensures that the project.yaml file exists if it's supposed to.
+     * Ensures that the project config YAML files exist if they’re supposed to
      *
      * @since 3.5.0
      */
     public static function ensureProjectConfigFileExists()
     {
-        if (!file_exists(Craft::$app->getPath()->getProjectConfigFilePath())) {
-            $projectConfig = Craft::$app->getProjectConfig();
+        $projectConfig = Craft::$app->getProjectConfig();
+
+        if ($projectConfig->writeYamlAutomatically && !$projectConfig->getDoesYamlExist()) {
             static::stdout('Generating project config files from the loaded project config ... ', static::FG_YELLOW);
             $projectConfig->regenerateYamlFromConfig();
             static::stdout('done' . PHP_EOL, static::FG_GREEN);